==> step5_visualization.qmd <==
# Advanced visualizations
https://hbctraining.github.io/Intro-to-DGE/lessons/06_DGE_visualizing_results.html

```{r}
# load libraries
library(BiocManager)
library(DESeq2)
library(tidyverse)
library(RColorBrewer)
library(pheatmap)
library(DEGreport)
library(tximport)
library(ggplot2)
library(ggrepel)
```

```{r}
# Load objects from previous analysis
load("intermediate/deg_objects.RData")
tx2gene <- read.delim("intermediate/mouse_annotations.tsv", header = TRUE)
```

```{r}
bv2_mg_meta_tb <- meta %>% 
              rownames_to_column(var="samplename") %>% 
              as_tibble()
bv2_mg_meta_tb
```

```{r}
# DESeq2 creates a matrix when you use the counts() function
## First convert normalized_counts to a data frame and transfer the row names to a new column called "gene"
normalized_counts <- counts(dds, normalized=T) %>% 
                     data.frame() %>%
                     rownames_to_column(var="gene")
```

```{r}
# check tx2gene annotations
head(tx2gene)
```

```{r}
# Next, merge together (ensembl IDs) the normalized counts data frame with a subset of the annotations in the tx2gene data frame (only the columns for ensembl gene IDs and gene symbols)
ens2gene <- tx2gene %>% 
               dplyr::select(gene_id, gene_name) %>% 
               dplyr::distinct()
head(ens2gene)
nrow(ens2gene)
```

```{r}
# check the normalized counts data frame
head(normalized_counts)
```

```{r}
## This will bring in a column of gene symbols
normalized_counts <- merge(normalized_counts, ens2gene, by.x="gene", by.y="gene_id")
head(normalized_counts)
```

```{r}
# Now create a tibble for the normalized counts
normalized_counts_tb <- normalized_counts %>%
                     as_tibble()
head(normalized_counts_tb)
```

```{r}
# Find the Ensembl ID
ens2gene[ens2gene$gene_name == "Axl", "gene_id"]
```

```{r}
# Save plotcounts to a data frame object
d_axl <- plotCounts(dds, gene="ENSMUSG00000002602", intgroup="sampletype", returnData=TRUE)

# Plot the normalized counts
ggplot(d_axl, aes(x = sampletype, y = count, color = sampletype)) + 
    geom_point(position=position_jitter(w = 0.1,h = 0)) +
    # geom_text_repel(aes(label = rownames(d_axl))) + 
    # theme_bw() +
    ggtitle("Axl") +
    theme(plot.title = element_text(hjust = 0.5))
```

```{r}
# the code block above only creates the plot in the R session, but doesn't save it to a file. To save it, we need to assign the plot to an object and then use ggsave() to export it.

# Create the plot object
p_axl <- ggplot(d_axl, aes(x = sampletype, y = count, color = sampletype)) + 
    geom_point(position=position_jitter(w = 0.1,h = 0)) +
    ggtitle("Axl") +
    theme(plot.title = element_text(hjust = 0.5))

# Export to PNG and SVG
ggsave("plots/plot12_axl_norm_counts.png", plot = p_axl, width = 5, height = 4)
ggsave("plots/plot12_axl_norm_counts.svg", plot = p_axl, width = 5, height = 4)
```

```{r}
# check significant results table
head(sig_tb)
# check normalized counts table
head(normalized_counts_tb)
```

```{r}
### Extract normalized expression for significant genes
norm_sig_tb <- normalized_counts_tb[,c(1:10)] %>% 
              dplyr::filter(gene %in% sig_tb$gene)  
head(norm_sig_tb)
nrow(norm_sig_tb)
```

```{r}
# check metadata df
meta
```

```{r}
### Set a color palette
heat_colors <- brewer.pal(6, "YlOrRd")

### Run pheatmap using the metadata data frame for the annotation
pheatmap(norm_sig_tb[2:9], 
    color = heat_colors, 
    cluster_rows = T, 
    show_rownames = F,
    annotation = meta, 
    border_color = NA, 
    fontsize = 10, 
    scale = "row", 
    # fontsize_row = 10, 
    # height = 20
    )
```

```{r}
sig_heatmap <- pheatmap(norm_sig_tb[2:9], 
    color = heat_colors, 
    cluster_rows = T, 
    show_rownames = F,
    annotation = meta, 
    border_color = NA, 
    fontsize = 10, 
    scale = "row", 
    )

# Save the heatmap to a file
ggsave("plots/plot13_sig_genes_heatmap.png", plot = sig_heatmap, width = 6, height = 8)
ggsave("plots/plot13_sig_genes_heatmap.svg", plot = sig_heatmap, width = 6, height = 8)
```

```{r}
## Obtain logical vector where TRUE values denote padj values < 0.05 and log2 fold change >= 2 in either direction
res_table_shrunken_label_tb <- res_table_shrunken_tb %>% 
                  dplyr::mutate(threshold = padj < 0.05 & abs(log2FoldChange) >= 2)
head(res_table_shrunken_label_tb)
```

```{r}
## Volcano plot
ggplot(res_table_shrunken_label_tb) +
    geom_point(aes(x = log2FoldChange, y = -log10(padj), colour = threshold), 
        size=1, alpha=0.5) +
    # ggtitle("Mov10 overexpression") +
    xlab("log2 fold change") + 
    ylab("-log10 adjusted p-value") +
    #scale_y_continuous(limits = c(0,50)) +
    theme(legend.position = "none",
          plot.title = element_text(size = rel(1.5), hjust = 0.5),
          axis.title = element_text(size = rel(1.25)))  
```

```{r}
## Add all the gene symbols as a column
res_table_shrunken_label_gene_tb <- bind_cols(res_table_shrunken_label_tb, 
                    symbol=ens2gene$gene_name[match(res_table_shrunken_label_tb$gene, ens2gene$gene_id)]
                    )
head(res_table_shrunken_label_gene_tb)
```

```{r}
## Create an empty column to indicate which genes to label
res_table_shrunken_label_gene_tb <- res_table_shrunken_label_gene_tb %>% dplyr::mutate(genelabels = "")
head(res_table_shrunken_label_gene_tb)
```

```{r}
## Sort by padj values 
res_table_shrunken_label_gene_tb <- res_table_shrunken_label_gene_tb %>% dplyr::arrange(padj)
head(res_table_shrunken_label_gene_tb)
```

```{r}
# check from which gene the padj values are not 0
# View(res_table_shrunken_label_gene_tb)
```

```{r}
## Populate the genelabels column with contents of the gene_name column
res_table_shrunken_label_gene_tb$genelabels[1:100] <- as.character(res_table_shrunken_label_gene_tb$symbol[1:100])
head(res_table_shrunken_label_gene_tb)
```

```{r}
ggplot(res_table_shrunken_label_gene_tb, aes(x = log2FoldChange, y = -log10(padj))) +
    geom_point(aes(colour = threshold), size=1, alpha=0.5) +
    geom_text_repel(aes(label = genelabels)) +
    # ggtitle("Mov10 overexpression") +
    xlab("log2 fold change") + 
    ylab("-log10 adjusted p-value") +
    theme(legend.position = "none",
          plot.title = element_text(size = rel(1.5), hjust = 0.5),
          axis.title = element_text(size = rel(1.25))) 
```

```{r}
# Find the Axl row
res_table_shrunken_label_gene_tb[res_table_shrunken_label_gene_tb$symbol == "Axl", ]
```

```{r}
# Save objects for functional analysis
save(res_table_shrunken_label_gene_tb, ens2gene, file = "intermediate/step5_objects.RData")
```
